requirements.txt
# ─── Core Framework ───
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
python-multipart>=0.0.18
pydantic>=2.10,<3.0
pydantic-settings>=2.7,<3.0

# ─── LLM ───
google-generativeai>=0.8,<1.0

# ─── Embeddings & Vector Store ───
sentence-transformers>=3.0,<4.0
faiss-cpu>=1.8,<2.0
numpy>=1.26,<2.0

# ─── Document Processing ───
pypdf>=5.0,<6.0
langchain-text-splitters>=0.3,<1.0

# ─── Web Search ───
httpx>=0.28,<1.0

# ─── Utilities ───
python-dotenv>=1.0,<2.0
loguru>=0.7,<1.0
